refactor UILabel -> BlinkingLabel
