+++ Sources/Controller/FirstPageViewController.swift
@@ -8,24 +8,12 @@
 
 import UIKit
 
-extension UILabel {
-    func blink() {
-        self.alpha = 0.0;
-        UIView.animate(withDuration: 0.9, //Time of duration
-            delay: 0.0,
-            options: [.autoreverse, .repeat],
-            animations: { [weak self] in self?.alpha = 1.0 },
-            completion: { [weak self] _ in self?.alpha = 0.0 })
-    }
-}
-
 class FirstPageViewController: UIViewController {
-    
     
     @IBAction func oneEuroStartButton(_ sender: UIButton) {
     }
     
-    @IBOutlet weak var blinkingLabel: UILabel!
+    @IBOutlet weak var blinkingLabel: BlinkingLabel!
 
     override func viewDidLoad() {
         super.viewDidLoad()
